pyproject.toml
[project]
name = "taxonomy_project"
version = "0.1.0"
description = "Word taxonomy manager"
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "nltk",
    "typing-extensions",
    "openai",
    "requests",
    "python-dotenv",
    "anthropic>=0.43.1",
    "ollama>=0.4.6",
    "rich>=13.9.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
taxonomy = "src.main:cli"
